requirements.txt
PyMySQL==0.9.2
Django==2.0.4
PyJWT==1.6.4
djangorestframework==3.9.0
